requirements.txt
pyside2>=5.15.2
pillow
briefcase